Copy editing (#1)
* Copy editing

Fix a few little typos and clean up the READMEs for someone who isn't
the original dev. Also, a bit more explanatory direction for n00bs.

* Code tour

Add a tour of the code to the README so that n00bs can understand
what's going on.

* Add foreman installation to README

Had previously forgotten foreman as a tool that needed to be
installed.